Use distribute to package application for further deployment
