+++ README.md
@@ -1,2 +1,15 @@
 # daimaduan.com
 source code of daimaduan.com
+
+## 项目依赖
+
+* Python 2.6+
+* pip 6+
+* mongodb
+
+## 本地运行开发服务器
+
+1. 启动MongoDB
+2. `cp daimaduan/config.cfg.example daimaduan/config.cfg`
+3. `python setup.py develop`
+4. `cd daimaduan && python runserver.py`
