+++ daimaduan/__init__.py
@@ -1,0 +1,13 @@
+from beaker.middleware import SessionMiddleware
+
+from daimaduan.bootstrap import app
+import daimaduan.views
+
+
+session_opts = {
+    'session.type': 'file',
+    'session.cookie_expires': 7 * 24 * 3600,
+    'session.data_dir': './sessions',
+    'session.auto': True
+}
+application = SessionMiddleware(app, session_opts)
