+++ daimaduan/config.cfg.sample
