+++ daimaduan/templates/pastes/view.html
